+++ backend/src/main/java/org/prgrms/devconnect/domain/define/Timestamp.java
@@ -1,0 +1,20 @@
+package org.prgrms.devconnect.domain.define;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+import lombok.Getter;
+import org.prgrms.devconnect.domain.define.CreateTimestamp;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+@Getter
+public abstract class Timestamp extends CreateTimestamp {
+
+  @LastModifiedDate
+  @Column(name = "updated_at")
+  private LocalDateTime updatedAt;
+}
