+++ backend/src/main/java/org/prgrms/devconnect/domain/define/chatting/entity/constant/ChattingRoomStatus.java
@@ -1,0 +1,13 @@
+package org.prgrms.devconnect.domain.define.chatting.entity.constant;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum ChattingRoomStatus {
+  ACTIVE("활성화"),
+  INACTIVE("비활성화");
+
+  private final String text;
+}
