+++ backend/src/main/java/org/prgrms/devconnect/domain/define/jobpost/entity/JobPostTechStackMapping.java
@@ -1,0 +1,36 @@
+package org.prgrms.devconnect.domain.define.jobpost.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.prgrms.devconnect.domain.define.techstack.entity.TechStack;
+
+@Entity
+@Table(name = "job_post_tech_stack_mapping")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class JobPostTechStackMapping {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "id")
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "job_post_id", nullable = false)
+  private JobPost jobPost;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "tech_stack_id", nullable = false)
+  private TechStack techStack;
+
+}
