+++ backend/src/main/java/org/prgrms/devconnect/domain/define/jobpost/entity/constant/JobType.java
@@ -1,0 +1,17 @@
+package org.prgrms.devconnect.domain.define.jobpost.entity.constant;
+
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum JobType {
+
+  REGULAR("정규직"),
+  IRREGULAR("비정규직"),
+  CONTRACT("인턴");
+
+  private final String text;
+
+}
