+++ backend/src/main/java/org/prgrms/devconnect/domain/define/jobpost/entity/constant/Status.java
@@ -1,0 +1,16 @@
+package org.prgrms.devconnect.domain.define.jobpost.entity.constant;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+
+@Getter
+@RequiredArgsConstructor
+public enum Status {
+
+  RECRUITING("모집중"),
+  DEADLINE("마감"),
+  DELETED("삭제됨");
+
+  private final String text;
+}
