+++ backend/src/main/java/org/prgrms/devconnect/domain/define/member/entity/Member.java
@@ -1,0 +1,61 @@
+package org.prgrms.devconnect.domain.define.member.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.prgrms.devconnect.domain.define.Timestamp;
+import org.prgrms.devconnect.domain.define.member.entity.constant.Interest;
+
+@Entity
+@Table(name = "member")
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Member extends Timestamp {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "member_id")
+  private Long memberId;
+
+  @Column(name = "email", length = 100)
+  private String email;
+
+  @Column(name = "password", length = 100)
+  private String password;
+
+  @Column(name = "nickname", length = 50)
+  private String nickname;
+
+  @Column(name = "affiliation", length = 100)
+  private String affiliation;
+
+  @Column(name = "career")
+  private int career;
+
+  @OneToMany(mappedBy = "member")
+  private List<MemberTechStackMapping> memberTechStacks = new ArrayList<>();
+
+  @Column(name = "self_introduction")
+  private String selfIntroduction;
+
+  @Column(name = "blog_link", length = 100)
+  private String blogLink;
+
+  @Column(name = "github_link", length = 100)
+  private String githubLink;
+
+  @Enumerated(value = EnumType.STRING)
+  @Column(name = "interest", length = 100)
+  private Interest interest;
+}
