+++ backend/src/main/java/org/prgrms/devconnect/domain/define/member/entity/constant/Interest.java
@@ -1,0 +1,15 @@
+package org.prgrms.devconnect.domain.define.member.entity.constant;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Interest {
+
+  STUDY("스터디"),
+  TEAM_PROJECT("팀 프로젝트"),
+  JOBPOST("채용 공고");
+
+  private final String text;
+}
