--- backend/src/main/java/org/prgrms/devconnect/member/entity/favoriteboard/FavoriteBoard.java
@@ -1,36 +1,0 @@
-package org.prgrms.devconnect.member.entity.favoriteboard;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.prgrms.devconnect.board.entity.Board;
-import org.prgrms.devconnect.member.entity.Member;
-
-@Entity
-@Table(name = "favorite_board")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class FavoriteBoard {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "favorite_board_id")
-  private Long favoriteBoardId;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "member_id", nullable = false)
-  private Member member;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "board_id", nullable = false)
-  private Board board;
-}
